refactor(search): read fs.search on this Mac too
The Search pane ran two matchers. A checkout on this Mac was searched by
`ContentSearch` shelling out to `git grep`; a checkout on a device was
searched by the daemon's `fs.search`, which walks with the `ignore` crate
and matches with `aho-corasick`. The two disagree by construction — the
daemon deliberately searches non-repo roots and walks nested repos — so
one query in one pane returned different hits depending on which box held
the code.

Local is not a special case: it is the device whose route is a Unix
socket. `.thisMac` now carries a `DeviceFileProvider` on `.local`, the
same plane the file tree has read every checkout through since #561, and
`find` has one body instead of one per machine.

`SearchScope` stays. It gates more than the engine — the Finder drag, the
editor-versus-preview open, the placeholder — and those are real client
differences. Only the duplicated matcher is gone.

`SearchExcerptTests` covered client-side windowing and span computation.
That work is host-side now and `files.rs` already pins it:
`spans_follow_the_smart_case_rule`, `a_long_line_is_windowed_around_its_hit`.

Cancel-on-keystroke still reaches the worker — the abandoned call cancels
its `fs.search` by request id and the daemon drops the walk.
